refactor: move data source to the view model
